print number

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -8,7 +8,7 @@
 void print_number(int n)
 {
 int a, b = 0, i = 0, x, y, z;
-double d = 1;
+int d = 1;
 if (n < 0)
 {
 n = n * (-1);
